requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM and AI
groq>=0.4.0
instructor>=0.4.0

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Text processing
tiktoken>=0.5.0
nltk>=3.8.0

# Retry logic
tenacity>=8.2.0

# Logging (usually included in Python, but listed for completeness)
# logging - built-in

